Split the code between RegexApp and RegexLabel.
What was the old RegexApp is now RegexLabel (this to avoid problems with 
the touch being held captive by the Main Layout, not being able to 
activate the text input for example).

Now the scrolling function is managed at the label level, enabling the 
textinput label to get its touches outside of the label.